+++ src/sql/plugins/stock/modules/recommendation/plugin/create_db_stock_recommendation.sql
@@ -5,7 +5,7 @@
 CREATE TABLE IF NOT EXISTS stock_recommendation (
   username varchar(255) NOT NULL,
   id_product int(11) NOT NULL,
-  score FLOAT NOT NULL,
+  score DOUBLE NOT NULL,
   PRIMARY KEY (username , id_product ),
   INDEX( username )
 ) ENGINE=InnoDB  DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
